fix: seperate auth and user helpers
